+++ HISTORY.md
@@ -29,6 +29,7 @@
 * Ignore and log timezone errors when attempting to reuse responses cached in `requests-cache <= 1.1`
 * Fix error handling with `stale_if_error` during revalidation requests
 * By default, do not automatically close backend connections when using `install_cache()`
+* Fix request headers sent when `expire_after` is set to `DO_NOT_CACHE`
 
 ### 1.2.1 (2024-06-18)